Move change-password & securityQs to components
