+++ apps/user-portal/src/components/index.ts
@@ -16,8 +16,10 @@
  * under the License.
  */
 
-export * from "./ui";
+export * from "./auth-context";
+export * from "./change-password";
 export * from "./header";
 export * from "./notification";
-export * from "./auth-context";
 export * from "./protected-route";
+export * from "./security-qs";
+export * from "./ui";
